Move PathMapping and ProcessorMapping out of native app logic
