+++ src/snowflake/cli/_plugins/nativeapp/codegen/compiler.py
@@ -37,9 +37,7 @@ from snowflake.cli._plugins.nativeapp.feature_flags import FeatureFlag
 from snowflake.cli.api.cli_global_context import get_cli_context
 from snowflake.cli.api.console import cli_console as cc
 from snowflake.cli.api.metrics import CLICounterField
-from snowflake.cli.api.project.schemas.v1.native_app.path_mapping import (
-    ProcessorMapping,
-)
+from snowflake.cli.api.project.schemas.entities.common import ProcessorMapping
 
 SNOWPARK_PROCESSOR = "snowpark"
 NA_SETUP_PROCESSOR = "native app setup"
