+++ tests/project/test_config.py
@@ -20,10 +20,10 @@ from unittest import mock
 import pytest
 from snowflake.cli.api.project.definition import load_project
 from snowflake.cli.api.project.errors import SchemaValidationError
+from snowflake.cli.api.project.schemas.entities.common import PathMapping
 from snowflake.cli.api.project.schemas.project_definition import (
     build_project_definition,
 )
-from snowflake.cli.api.project.schemas.v1.native_app.path_mapping import PathMapping
 
 from tests.nativeapp.factories import PdfV10Factory
