+++ tests/project/test_project_definition_v2.py
@@ -15,7 +15,6 @@ from pathlib import Path
 
 import pytest
 from snowflake.cli._plugins.snowpark.snowpark_entity_model import (
-    PathMapping,
     SnowparkEntityModel,
 )
 from snowflake.cli.api.project.definition_conversion import (
@@ -23,6 +22,7 @@ from snowflake.cli.api.project.definition_conversion import (
 )
 from snowflake.cli.api.project.definition_manager import DefinitionManager
 from snowflake.cli.api.project.errors import SchemaValidationError
+from snowflake.cli.api.project.schemas.entities.common import PathMapping
 from snowflake.cli.api.project.schemas.entities.entities import (
     ALL_ENTITIES,
     ALL_ENTITY_MODELS,
